Changed types

+++ db/migrations/20250409081412_refactoring.tx.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `volunteer` (
-    `tg_id` int(11) NOT NULL,
+    `tg_id` BIGINT NOT NULL,
     `first_name` varchar(255) NOT NULL,
     `last_name` varchar(255) NOT NULL,
     `middle_name` varchar(255) NOT NULL,
